refactor: 예약 메일 수정 방식 개선
